EditStyleDialog: remove redundant null checks
Showing EditStyleDialog with m_settingsWidget == nullptr makes no sense,
because the style cannot be edited. This dialog is created only in
SourceFormatterSelectionEdit::editStyle() and only when
ISourceFormatter::hasEditStyleWidget() returns true. Furthermore, both
AStylePlugin::editStyleWidget() and
CustomScriptPlugin::editStyleWidget() return valid pointers. Make this
fact a precondition of EditStyleDialog's constructor to simplify the
implementation of this class.