Create editStyleWidget only before showing it
SourceFormatterSelectionEdit::enableStyleButtons() calls
ISourceFormatter::editStyleWidget(), checks if it is null and
immediately destroys the widget. This occurs every time the user selects
a language, formatter or style.

SourceFormatterSelectionEdit::editStyle() also creates the widget just
to compare it with nullptr, but at least this function is called much
less often.

This is a lot of extra work => introduce
ISourceFormatter::hasEditStyleWidget() to improve performance.