TestCodeCompletion: fix mistakes in XFAIL test data
The name of a function parameter cannot equal a C++ keyword "double".

"instance.doubleItem" and "pInstance->doubleItem" are still missing from
the actual completion item list for this data row. The second "function"
string is in the actual list, but missing from the expected list. So I
added it under the same `#if` as in the previous test data row.