Port uses item repository to non-recursive mutex
Don't lock the mutex internally but instead lock it externally
before accessing the repository. This will mean the same lock is
held for longer periods of times instead of getting relocked
multiple times, but I believe that the old code was not actually
safe in that regard - what guaranteed that itemFromIndex returned
something isn't getting removed while we are accessing it?