Always assume external mutex locking for ItemRepository
Instead, take the mutex type which allows us to use either a
QRecursiveMutex or a normal QMutex. Then simplify the code
and remove the ItemRepositoryPrivate again, which is now obsolete.

Undo the changes we did previously for ItemRequest::destroy and
operate on AbstractItemRepository& there again instead of a template
parameter for the repo.