Make ItemRepository accesses in DUChain code mutex safe
I.e. port them over to the new LockedItemRepository::op paradigm.
This is all pretty straight forward, but clearly shows one issue
that has been lurking in the code base for ages - in loadInformation
we get an item ptr from the repo and then access it after unlocking
the mutex which is not safe at all imo. We'll have to revisit this
in the future.