Eliminate ItemRepositoryRegistryPrivate::open()
This function was called only from ItemRepositoryRegistry's constructor.
Move its implementation to ItemRepositoryRegistryPrivate() to make this
clear. Remove ItemRepositoryRegistryPrivate::m_mutex lock from
ItemRepositoryRegistryPrivate() because a constructor of a class doesn't
need to lock its own mutex.

Remove the `m_path == path` check from the constructor as at that point
m_path is empty and path is asserted to be non-empty.

Remove the code that opens contained repositories from the constructor
as m_repositories is always empty there.

Remove ItemRepositoryRegistryPrivate::m_owner. This data member is now
used only in the constructor, so there is no need to store it.